#1544: Extract BrushFaceReference and use it in BrushFaceSnapshot as well.
