+++ common/src/Model/BrushFaceReference.cpp
@@ -17,27 +17,23 @@
  along with TrenchBroom. If not, see <http://www.gnu.org/licenses/>.
  */
 
-#include "BrushFaceSnapshot.h"
+#include "BrushFaceReference.h"
 
 #include "Model/Brush.h"
+#include "Model/BrushFace.h"
 
 namespace TrenchBroom {
     namespace Model {
-        BrushFaceSnapshot::BrushFaceSnapshot(BrushFace* face, TexCoordSystem* coordSystem) :
-        m_faceRef(face),
-        m_attribs(face->attribs().takeSnapshot()),
-        m_coordSystemSnapshot(coordSystem->takeSnapshot()) {}
-        
-        BrushFaceSnapshot::~BrushFaceSnapshot() {
-            delete m_coordSystemSnapshot;
-            m_coordSystemSnapshot = nullptr;
+        BrushFaceReference::BrushFaceReference(Model::BrushFace* face) :
+        m_facePlane(face->boundary()),
+        m_brush(face->brush()) {
+            ensure(m_brush != nullptr, "face without a brush");
         }
 
-        void BrushFaceSnapshot::restore() {
-            BrushFace* face = m_faceRef.resolve();
-            face->setAttribs(m_attribs);
-            if (m_coordSystemSnapshot != nullptr)
-                face->restoreTexCoordSystemSnapshot(m_coordSystemSnapshot);
+        Model::BrushFace* BrushFaceReference::resolve() const {
+            Model::BrushFace* face = m_brush->findFace(m_facePlane);
+            ensure(face != nullptr, "couldn't find face");
+            return face;
         }
     }
 }
